+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### 0.2.0
 
+- You can now search for a mob and click it to see everything it drops
 - Added regex support to advanced search
 - Added missing lang entries